refactor(wallet): refactored some of core services to wallet
