+++ honestcash-v2/src/app/story/components/unlock-button/unlock-button.component.html
@@ -1,6 +1,6 @@
-<shared-transaction-button
+<wallet-transaction-button
   [story]="story"
   [actionText]="'Unlock'"
   [actionLoadingText]="'Unlocking...'"
   [onTransactionComplete]="onTransactionComplete"
-  [transactionType]="TRANSACTION_TYPES.Unlock"></shared-transaction-button>
+  [transactionType]="TRANSACTION_TYPES.Unlock"></wallet-transaction-button>
