+++ honestcash-v2/src/app/story/components/upvote-button/story-upvote-button.component.html
@@ -1,7 +1,7 @@
-<shared-transaction-button
+<wallet-transaction-button
   [story]="story"
   [actionText]="'Upvote'"
   [actionLoadingText]="'Upvoting...'"
   [onTransactionComplete]="onTransactionComplete"
   [transactionType]="TRANSACTION_TYPES.Upvote"
-  [isSmallButton]="isSmallButton"></shared-transaction-button>
+  [isSmallButton]="isSmallButton"></wallet-transaction-button>
