+++ honestcash-v2/src/app/story/components/upvotes/story-upvotes.component.ts
@@ -5,7 +5,7 @@ import {Observable, Subscription} from 'rxjs';
 import {StoryState} from '../../store/story.state';
 import {Store} from '@ngrx/store';
 import {AppStates, selectStoryState} from '../../../app.states';
-import {TRANSACTION_TYPES} from '../../../../core/shared/models/transaction';
+import {TRANSACTION_TYPES} from '../../../wallet/models/transaction';
 
 @Component({
   selector: 'story-upvotes',
