+++ honestcash-v2/src/app/story/story.module.ts
@@ -23,6 +23,7 @@ import {StoryUnlockButtonComponent} from './components/unlock-button/unlock-butt
 import {StoryCommentCardActionsComponent} from './components/comment-card-actions/comment-card-actions.component';
 import {StoryCommentCardBodyComponent} from './components/comment-card-body/story-comment-card-body.component';
 import {StoryCommentCardHeaderComponent} from './components/comment-card-header/story-comment-card-header.component';
+import {WalletSharedModule} from '../wallet/wallet-shared.module';
 
 @NgModule({
   declarations: [
@@ -49,6 +50,7 @@ import {StoryCommentCardHeaderComponent} from './components/comment-card-header/
   imports: [
     StoryRoutingModule,
     StorySharedModule,
+    WalletSharedModule,
     CommonModule,
     SharedModule,
   ],
