+++ honestcash-v2/src/app/wallet/components/receipt/receipt.component.scss
