+++ honestcash-v2/src/app/wallet/components/receipt/receipt.component.ts
@@ -1,14 +1,14 @@
 import {Component, OnInit} from '@angular/core';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 import {ITransaction, TRANSACTION_TYPES} from '../../models/transaction';
-import Story from '../../../../app/story/models/story';
+import Story from '../../../story/models/story';
 
 @Component({
-  selector: 'shared-receipt',
+  selector: 'wallet-receipt',
   templateUrl: './receipt.component.html',
   styleUrls: ['./receipt.component.scss']
 })
-export class SharedReceiptComponent implements OnInit {
+export class WalletReceiptComponent implements OnInit {
 
   public transactionType: TRANSACTION_TYPES;
   public TRANSACTION_TYPES = TRANSACTION_TYPES;
