+++ honestcash-v2/src/app/wallet/components/transaction-button/transaction-button.component.scss
