+++ honestcash-v2/src/app/wallet/services/upvote.transaction.strategy.ts
