+++ honestcash-v2/src/app/wallet/wallet-shared.module.ts
@@ -1,16 +1,25 @@
 import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
 import {WalletBalanceComponent} from './components/balance/balance.component';
+import {WalletReceiptComponent} from './components/receipt/receipt.component';
+import {WalletTransactionButtonComponent} from './components/transaction-button/transaction-button.component';
+import {SharedModule} from '../../core/shared.module';
 
 @NgModule({
   declarations: [
     WalletBalanceComponent,
+    WalletReceiptComponent,
+    WalletTransactionButtonComponent,
   ],
   imports: [
-    CommonModule,
+    SharedModule,
+  ],
+  entryComponents: [
+    WalletReceiptComponent,
   ],
   exports: [
     WalletBalanceComponent,
+    WalletReceiptComponent,
+    WalletTransactionButtonComponent,
   ],
 })
 export class WalletSharedModule {
