+++ honestcash-v2/src/core/shared.module.ts
@@ -8,11 +8,7 @@ import {SharedHeaderComponent} from './shared/components/header/header.component
 import {ShardHeaderProfileMenuComponent} from './shared/components/header-profile-menu/header-profile-menu.component';
 import {SharedNotFoundComponent} from './shared/pages/not-found/not-found.component';
 import {SharedLoadingIndicatorComponent} from './shared/components/loading-indicator/loading-indicator.component';
-import {WalletSharedModule} from '../app/wallet/wallet-shared.module';
-import {SharedReceiptComponent} from './shared/components/receipt/receipt.component';
 import {QRCodeModule} from 'angularx-qrcode';
-import {SharedTransactionButtonComponent} from './shared/components/transaction-button/transaction-button.component';
-import {TransactionService} from './shared/services/transaction.service';
 import {WalletService} from '../app/wallet/services/wallet.service';
 import {StoryService} from '../app/story/services/story.service';
 import {ToastrModule} from 'ngx-toastr';
@@ -26,25 +22,14 @@ import {RouterModule} from '@angular/router';
     SharedHonestLogoComponent,
     SharedLoadingIndicatorComponent,
     SharedNotFoundComponent,
-    SharedReceiptComponent,
-    SharedTransactionButtonComponent,
   ],
   imports: [
     CommonModule,
     FontAwesomeModule,
     NgbModule,
-    WalletSharedModule,
     QRCodeModule,
     ToastrModule.forRoot(),
     RouterModule,
-  ],
-  providers: [
-    TransactionService,
-    WalletService,
-    StoryService,
-  ],
-  entryComponents: [
-    SharedReceiptComponent,
   ],
   exports: [
     CommonModule,
@@ -55,8 +40,6 @@ import {RouterModule} from '@angular/router';
     ShardHeaderProfileMenuComponent,
     SharedHonestLogoComponent,
     SharedLoadingIndicatorComponent,
-    SharedReceiptComponent,
-    SharedTransactionButtonComponent,
   ]
 })
 export class SharedModule {
