+++ honestcash-v2/src/core/store/core.effects.ts
@@ -1,13 +1,12 @@
 import {Inject, Injectable, PLATFORM_ID} from '@angular/core';
 import {Actions, Effect, ofType} from '@ngrx/effects';
 import {Observable} from 'rxjs';
-import {mergeMap, tap} from 'rxjs/operators';
+import {mergeMap} from 'rxjs/operators';
 import {UserSetup} from '../../app/user/store/user.actions';
 import {WalletSetup} from '../../app/wallet/store/wallet.actions';
 import {CoreActionTypes} from './core.actions';
 import {LocalStorageToken} from '../shared/helpers/local-storage.helper';
 import {AuthSetup} from '../../app/auth/store/auth.actions';
-import {WalletService} from '../../app/wallet/services/wallet.service';
 
 
 @Injectable()
